aula_02_0/requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.16.1
PyOpenGL>=3.1.0
PyOpenGL_accelerate>=3.1.0
pygame>=2.1.0